Variable array API
This commit adds a new representation to Dr.Jit that groups a number of
regular variables into a "variable array". The API exposes convenient
functionality to read and write elements of this variable array.
Pre-initialized arrays storing a literal constant can also be created.

The main difference between variable arrays and a collection of regular
variables is that this representation supports reads/writes using
computed indices. This enables previously impossible use cases of Dr.Jit.

At an implementation level, array variables are stored on the
stack---for example, the LLVM version uses ``alloca()`` to reserve
memory within the function scope.  That said, it is possible that
optimization passes in the backend compiler might turn this
stack-allocated memory back to CPU/GPU registers.