Packet gather/scatter operations
Vector gather and scatter are conceptually neat operations, but
unfortunately CPUs aren't very good at executing them efficiently in
practice. In fact, their performance has decreased over time due to the
mitigation of the "Downfall" side-channel vulnerability
(https://downfall.page) that disables speculation involving gathers.

Often, we're gathering/scattering not just one value but several
adjacent ones. In this case, it can be far more efficient to combine
regular packet memory operations with a transpose that brings them into
the right order (this assumes that the operation accesses an aligned
region, whose size is a power-of-two size).

This commit adds a specialized compilation strategy that realizes this
optimization on the LLVM and CUDA backends. On the LLVM backend, it
leads to dramatic (> 20x) speedups. Packet operations also improve
performance on the CUDA backend, where the improvement is smaller (on
the order of a few tens of percent) but still more than worthwhile.